+++ src/euphorie/content/browser/export.py
@@ -10,6 +10,7 @@ View name: @@export
 from Acquisition import aq_inner
 from Acquisition import aq_parent
 from euphorie.client.utils import HasText
+from euphorie.content import MessageFactory as _
 from euphorie.content.behaviors.toolcategory import IToolCategory
 from euphorie.content.browser.upload import COMMA_REPLACEMENT
 from euphorie.content.browser.upload import NSMAP
@@ -23,10 +24,12 @@ from euphorie.content.survey import get_tool_type
 from euphorie.content.utils import StripMarkup
 from euphorie.content.utils import StripUnwanted
 from lxml import etree
-from plone import api
+from plone.autoform.form import AutoExtensibleForm
 from Products.CMFPlone.utils import safe_bytes
-from Products.CMFPlone.utils import safe_nativestring
-from Products.Five import BrowserView
+from z3c.form import button
+from z3c.form import form
+from zope import schema
+from zope.interface import Interface
 
 
 try:
@@ -49,19 +52,137 @@ def getToken(field, value, default=None):
         return default
 
 
-class ExportSurvey(BrowserView):
-    """View name: @@export"""
+class IExportSurveySchema(Interface):
+    """Fields used for configuring the export"""
 
-    def __init__(self, context, request):
-        super(ExportSurvey, self).__init__(context, request)
-        self.use_existing_measures = api.portal.get_registry_record(
-            "euphorie.use_existing_measures", default=False
+    include_images = schema.Bool(
+        title=_("label_survey_export_include_images", default="Include images"),
+        description=_(
+            "description_survey_export_include_images",
+            default="All images contained in the OiRA tool can be exported along "
+            "with the textual content. When the OiRA tool is then imported again, "
+            "all images will be present. Please note that including images may "
+            "result in a considerably larger file size.",
+        ),
+        required=False,
+        default=False,
+    )
+    include_intro_text = schema.Bool(
+        title=_(
+            "label_survey_export_include_intro_text",
+            default="Include introduction text",
+        ),
+        description=_(
+            "description_survey_export_include_intro_text",
+            default="The introduction text describes the general purpose of the OiRA "
+            "tool. It usually gives the end users an idea what to expect in the tool "
+            "and whether it will be relevant for them. The default is to include "
+            "this text.",
+        ),
+        required=False,
+        default=True,
+    )
+    include_module_solution_texts = schema.Bool(
+        title=_(
+            "label_survey_export_include_module_solution_texts",
+            default="Include “Solution” texts on modules",
+        ),
+        description=_(
+            "description_survey_export_include_module_solution_texts",
+            "The “Solution” description” on modules is a deprecated kind of text "
+            "that is no longer being used. The default is to skip this.",
+        ),
+        required=False,
+        default=False,
+    )
+    include_module_description_texts = schema.Bool(
+        title=_(
+            "label_survey_export_include_module_description_texts",
+            default="Include “Description” texts on modules",
+        ),
+        description=_(
+            "description_survey_export_include_module_description_texts",
+            default="Every module comes with an introductive description. The default "
+            "is to include this text.",
+        ),
+        required=False,
+        default=True,
+    )
+    include_risk_legal_texts = schema.Bool(
+        title=_(
+            "label_survey_export_include_risk_legal_texts",
+            default="Include “Legal and policy references” texts on risks",
+        ),
+        description=_(
+            "description_survey_export_include_risk_legal_texts",
+            default="If the OiRA tool is exported because it should be made available "
+            "for a different country, the legal texts might not apply there. Therefore "
+            "the default is to skip these texts.",
+        ),
+        required=False,
+        default=True,
+    )
+    include_measures = schema.Bool(
+        title=_("label_survey_export_include_measures", default="Include measures"),
+        description=_(
+            "description_survey_export_include_measures",
+            default="For each risk, a number of common measures – sometimes also "
+            "referred to as “Solutions” – can be defined. While the default setting "
+            "is to include them in the export, they may be excluded if required, "
+            "for example to accommodate for a limited translation budget.",
+        ),
+        required=False,
+        default=True,
+    )
+
+
+class ExportSurvey(AutoExtensibleForm, form.Form):
+    """The upload view for a :obj:`euphorie.content.sector`
+
+    View name: @@upload
+    """
+
+    schema = IExportSurveySchema
+    ignoreContext = True
+    form_name = _(u"Export OiRA Tool")
+
+    @button.buttonAndHandler(_(u"Export"))
+    def handleExport(self, action):
+        (data, errors) = self.extractData()
+        self.include_images = data.get("include_images")
+        self.include_intro_text = data.get("include_intro_text")
+        self.include_module_solution_texts = data.get("include_module_solution_texts")
+        self.include_module_description_texts = data.get(
+            "include_module_description_texts"
         )
-        nsmap = NSMAP.copy()
-        del nsmap[None]
+        self.include_risk_legal_texts = data.get("include_risk_legal_texts")
+        self.include_measures = data.get("include_measures")
+        output = etree.Element("sector", nsmap=NSMAP)
+        self.exportSurvey(output, self.context)
+        response = self.request.response
+        filename = "%s.xml" % aq_parent(aq_inner(self.context)).id
+        response.setHeader(
+            "Content-Disposition", u'attachment; filename="%s"' % filename
+        )
+        response.setHeader("Content-Type", "text/xml")
+        self.request.response.write(
+            etree.tostring(
+                output, pretty_print=True, xml_declaration=True, encoding="utf-8"
+            )
+        )
+
+    def render(self):
+        if self.request.response.headers.get("content-disposition", "").startswith(
+            "attachment"
+        ):
+            # We just generated an attachment, no need to render the form
+            return ""
+        return super().render()
 
     def exportImage(self, parent, image, caption=None):
         """:returns: base64 encoded image."""
+        if not self.include_images:
+            return
         node = etree.SubElement(parent, "image")
         if image.contentType:
             node.attrib["content-type"] = image.contentType
@@ -81,7 +202,7 @@ class ExportSurvey(BrowserView):
         if getattr(survey, "external_id", None):
             node.attrib["external-id"] = survey.external_id
         etree.SubElement(node, "title").text = aq_parent(survey).title
-        if StripMarkup(survey.introduction):
+        if self.include_intro_text and StripMarkup(survey.introduction):
             etree.SubElement(node, "introduction").text = StripUnwanted(
                 survey.introduction
             )
@@ -159,13 +280,15 @@ class ExportSurvey(BrowserView):
         if getattr(module, "external_id", None):
             node.attrib["external-id"] = module.external_id
         etree.SubElement(node, "title").text = module.title
-        if HasText(module.description):
+        if self.include_module_description_texts and HasText(module.description):
             etree.SubElement(node, "description").text = StripUnwanted(
                 module.description
             )
         if module.optional:
             etree.SubElement(node, "question").text = module.question
-        if StripMarkup(module.solution_direction):
+        if self.include_module_solution_texts and StripMarkup(
+            module.solution_direction
+        ):
             etree.SubElement(node, "solution-direction").text = StripUnwanted(
                 module.solution_direction
             )
@@ -189,15 +312,13 @@ class ExportSurvey(BrowserView):
             risk.problem_description
         )
         etree.SubElement(node, "description").text = StripUnwanted(risk.description)
-        if StripMarkup(risk.legal_reference):
+        if self.include_risk_legal_texts and StripMarkup(risk.legal_reference):
             etree.SubElement(node, "legal-reference").text = StripUnwanted(
                 risk.legal_reference
             )
         etree.SubElement(node, "show-not-applicable").text = (
             "true" if risk.show_notapplicable else "false"
         )
-        if self.use_existing_measures:
-            etree.SubElement(node, "existing_measures").text = risk.existing_measures
         if risk.type == "risk":
             method = etree.SubElement(node, "evaluation-method")
             method.text = risk.evaluation_method
@@ -233,11 +354,14 @@ class ExportSurvey(BrowserView):
             if image is not None:
                 self.exportImage(node, image, getattr(risk, "caption" + postfix, None))
 
-        solutions = [child for child in risk.values() if ISolution.providedBy(child)]
-        if solutions:
-            sols = etree.SubElement(node, "solutions")
-            for solution in solutions:
-                self.exportSolution(sols, solution)
+        if self.include_measures:
+            solutions = [
+                child for child in risk.values() if ISolution.providedBy(child)
+            ]
+            if solutions:
+                sols = etree.SubElement(node, "solutions")
+                for solution in solutions:
+                    self.exportSolution(sols, solution)
         return node
 
     def exportSolution(self, parent, solution):
@@ -253,21 +377,3 @@ class ExportSurvey(BrowserView):
                 solution.requirements
             )
         return node
-
-    def __call__(self):
-        """:returns: an XML export of a Survey.
-        :rtype: str
-        """
-        output = etree.Element("sector", nsmap=NSMAP)
-        self.exportSurvey(output, self.context)
-        response = self.request.response
-        filename = "%s.xml" % aq_parent(aq_inner(self.context)).id
-        response.setHeader(
-            "Content-Disposition", u'attachment; filename="%s"' % filename
-        )
-        response.setHeader("Content-Type", "text/xml")
-        return safe_nativestring(
-            etree.tostring(
-                output, pretty_print=True, xml_declaration=True, encoding="utf-8"
-            )
-        )