+++ docs/changes.rst
@@ -8,6 +8,8 @@ Changelog
 - Consolidate the "survey view" in the CMS by merging back developments done in
   subprojects (OSHA-OiRA and Daimler).
   This includes the possibility to export the complete contents of an OiRA tool as Word file
+- For the XML export of Surveys, give the user the posibility of de-selecting
+  certain parts (images, legal texts, etc)
 
 
 14.0.1 (2022-03-17)